integramos los ficheros generateData t generateData1 en un único fichero
