+++ src/screens/Sales/SalesLineChart.js
@@ -1,10 +1,10 @@
 import Body from "components/layout/Body/Body";
 import {LineChart} from "../../components/composed/Charts/LineChart/LineChart";
-import {generateData, daysBetween} from "utils/generateData";
+import {generateDataLine, daysBetween} from "utils/generateData";
 
 const initialDate = "2012-01-01";
 const days = daysBetween(initialDate, "2016-12-31");
-const data = generateData(new Date(initialDate), days, [30, 80]);
+const data = generateDataLine(new Date(initialDate), days, [30, 80]);
 
 function SalesLineChart() {
 	return (
