+++ src/utils/generateData.js
@@ -1,9 +1,33 @@
+const returnNumFromRange = (min, max) => Math.floor(Math.random() * (max - min + 1) + min);
 const dayInMiliseconds = 24 * 60 * 60 * 1000;
-// const initialDate = new Date(2020, 4, 8);
-// const values = [30, 80];
 const randomValue = (min, max) => Math.floor(Math.random() * (max - min + 1) + min);
 
-export const generateData = (startDay, days, rangeValues) => {
+export const daysBetween = (firstDay, lastDay) => {
+	const fD = new Date(firstDay);
+	const lD = new Date(lastDay);
+	const miliSecondsDifference = Math.abs(fD - lD);
+	return Math.ceil(miliSecondsDifference / dayInMiliseconds) + 1;
+};
+
+export const generateData = (startDay, days, valueRange) => {
+	const data = [];
+
+	for (let i = 0; i < days; i++) {
+		// 24 * 60 * 60 * 1000 -> día en milisegundos
+		let day = new Date(startDay.getTime() + 24 * 60 * 60 * 1000 * i);
+		let option = {
+			day: day,
+			pisos: returnNumFromRange(valueRange[0], valueRange[1]),
+			garages: returnNumFromRange(valueRange[0], valueRange[1]),
+			locales: returnNumFromRange(valueRange[0], valueRange[1]),
+			chalets: returnNumFromRange(valueRange[0], valueRange[1]),
+		};
+		data.push(option);
+	}
+	return data;
+};
+
+export const generateDataLine = (startDay, days, rangeValues) => {
 	const data = [];
 	const [min, max] = rangeValues;
 
@@ -22,41 +46,61 @@ export const generateData = (startDay, days, rangeValues) => {
 	return data;
 };
 
-export const daysBetween = (firstDay, lastDay) => {
-	const fD = new Date(firstDay);
-	const lD = new Date(lastDay);
-	const miliSecondsDifference = Math.abs(fD - lD);
-	return Math.ceil(miliSecondsDifference / dayInMiliseconds) + 1;
-};
-
-export const groupByMonth = (yearlyData) => {
+export const groupByType = (yearlyData) => {
 	const months = {
-		0: 0,
-		1: 0,
-		2: 0,
-		3: 0,
-		4: 0,
-		5: 0,
-		6: 0,
-		7: 0,
-		8: 0,
-		9: 0,
-		10: 0,
-		11: 0,
+		0: {pisos: 0, locales: 0, garages: 0, chalets: 0},
+		1: {pisos: 0, locales: 0, garages: 0, chalets: 0},
+		2: {pisos: 0, locales: 0, garages: 0, chalets: 0},
+		3: {pisos: 0, locales: 0, garages: 0, chalets: 0},
+		4: {pisos: 0, locales: 0, garages: 0, chalets: 0},
+		5: {pisos: 0, locales: 0, garages: 0, chalets: 0},
+		6: {pisos: 0, locales: 0, garages: 0, chalets: 0},
+		7: {pisos: 0, locales: 0, garages: 0, chalets: 0},
+		8: {pisos: 0, locales: 0, garages: 0, chalets: 0},
+		9: {pisos: 0, locales: 0, garages: 0, chalets: 0},
+		10: {pisos: 0, locales: 0, garages: 0, chalets: 0},
+		11: {pisos: 0, locales: 0, garages: 0, chalets: 0},
 	};
 
 	for (let i = 0; i < yearlyData.length; i++) {
 		const el = yearlyData[i];
 		const curMonth = el.day.getMonth();
-		months[curMonth] = months[curMonth] += el.total;
+
+		months[curMonth].pisos = months[curMonth].pisos += el.pisos;
+		months[curMonth].locales = months[curMonth].locales += el.locales;
+		months[curMonth].garages = months[curMonth].garages += el.garages;
+		months[curMonth].chalets = months[curMonth].chalets += el.chalets;
 	}
 
-	let finalArr = [];
+	let totalpisos = [];
+	let totallocales = [];
+	let totalgarages = [];
+	let totalchalets = [];
+
 	for (const month in months) {
-		finalArr.push(months[month]);
+		totalpisos.push(months[month].pisos);
+		totallocales.push(months[month].locales);
+		totalgarages.push(months[month].garages);
+		totalchalets.push(months[month].chalets);
 	}
 
-	return finalArr;
+	return [{data: totalpisos}, {data: totallocales}, {data: totalgarages}, {data: totalchalets}];
+};
+
+export const groupByYear = (selectedYear, data) => {
+	return data.map((e) => e.day.getFullYear()).indexOf(parseInt(selectedYear));
+};
+
+export const getDaysInMonth = (month, year) => {
+	return new Date(year, parseInt(month) + 1, 0).getDate();
+};
+
+export const groupByFilter = (selectedMonth, selectedYear, data) => {
+	return data
+		.map((e) => {
+			return `${e.day.getMonth()}, ${e.day.getDate()}, ${e.day.getFullYear()}`;
+		})
+		.indexOf(`${selectedMonth}, 1, ${selectedYear}`);
 };
 
 export const getByDays = (monthlyData) => {
@@ -69,23 +113,24 @@ export const getByDays = (monthlyData) => {
 	return days;
 };
 
-const returnNumFromRange = (min, max) => Math.floor(Math.random() * (max - min + 1) + min);
-
-export const generateDataBar = (startDate, endDate, valueRange) => {
-	const data = [];
-	const days = daysBetween(startDate, endDate);
-	for (let i = 0; i < days; i++) {
-		let day = new Date(new Date(startDate).getTime() + dayInMiliseconds * i);
-		let option = {
-			day: day,
-			pisos: returnNumFromRange(valueRange[0], valueRange[1]),
-			garages: returnNumFromRange(valueRange[0], valueRange[1]),
-			locales: returnNumFromRange(valueRange[0], valueRange[1]),
-			chalets: returnNumFromRange(valueRange[0], valueRange[1]),
-		};
-		data.push(option);
+export const groupByTypeMonth = (monthData) => {
+	let totalpisos = 0;
+	let totallocales = 0;
+	let totalgarages = 0;
+	let totalchalets = 0;
+	for (let i = 0; i < monthData.length; i++) {
+		const el = monthData[i];
+		totalpisos += el.pisos;
+		totallocales += el.locales;
+		totalgarages += el.garages;
+		totalchalets += el.chalets;
 	}
-	return data;
+	return [
+		{data: [totalpisos]},
+		{data: [totallocales]},
+		{data: [totalgarages]},
+		{data: [totalchalets]},
+	];
 };
 
 export const groupByTypeYear = (yearlyData) => {
@@ -129,22 +174,32 @@ export const groupByTypeYear = (yearlyData) => {
 	return [{data: totalpisos}, {data: totallocales}, {data: totalgarages}, {data: totalchalets}];
 };
 
-export const groupByTypeMonth = (monthData) => {
-	let totalpisos = 0;
-	let totallocales = 0;
-	let totalgarages = 0;
-	let totalchalets = 0;
-	for (let i = 0; i < monthData.length; i++) {
-		const el = monthData[i];
-		totalpisos += el.pisos;
-		totallocales += el.locales;
-		totalgarages += el.garages;
-		totalchalets += el.chalets;
+export const groupByMonth = (yearlyData) => {
+	const months = {
+		0: 0,
+		1: 0,
+		2: 0,
+		3: 0,
+		4: 0,
+		5: 0,
+		6: 0,
+		7: 0,
+		8: 0,
+		9: 0,
+		10: 0,
+		11: 0,
+	};
+
+	for (let i = 0; i < yearlyData.length; i++) {
+		const el = yearlyData[i];
+		const curMonth = el.day.getMonth();
+		months[curMonth] = months[curMonth] += el.total;
 	}
-	return [
-		{data: [totalpisos]},
-		{data: [totallocales]},
-		{data: [totalgarages]},
-		{data: [totalchalets]},
-	];
+
+	let finalArr = [];
+	for (const month in months) {
+		finalArr.push(months[month]);
+	}
+
+	return finalArr;
 };
